refactor: add subfolder hooks/featureFlags
