+++ src/common/containers/FeatureGuard/index.tsx
@@ -1,6 +1,6 @@
 import { ReactNode } from 'react'
 
-import { useFeatureFlags } from 'common/hooks/useFeatureFlags'
+import { useFeatureFlags } from 'common/hooks/featureFlags/useFeatureFlags'
 
 interface FeatureGuardProps {
   featureFlag: string
