+++ src/common/hooks/featureFlags/useIsEthFlowBundlingEnabled.ts
@@ -1,6 +1,6 @@
 import { useIsBundlingSupported } from 'modules/wallet'
 
-import { useFeatureFlags } from 'common/hooks/useFeatureFlags'
+import { useFeatureFlags } from 'common/hooks/featureFlags/useFeatureFlags'
 
 export function useIsEthFlowBundlingEnabled(): boolean {
   const isBundlingSupported = useIsBundlingSupported()
