+++ src/common/hooks/featureFlags/useIsTxBundlingEnabled.ts
@@ -1,6 +1,6 @@
 import { useIsBundlingSupported } from 'modules/wallet'
 
-import { useFeatureFlags } from 'common/hooks/useFeatureFlags'
+import { useFeatureFlags } from 'common/hooks/featureFlags/useFeatureFlags'
 
 export function useIsTxBundlingEnabled(): boolean {
   const isBundlingSupported = useIsBundlingSupported()
