+++ src/modules/limitOrders/hooks/useHandleOrderPlacement.test.ts
@@ -9,7 +9,7 @@ import { safeBundleFlow } from 'modules/limitOrders/services/safeBundleFlow'
 import { tradeFlow } from 'modules/limitOrders/services/tradeFlow'
 import { TradeFlowContext } from 'modules/limitOrders/services/types'
 
-import { useIsTxBundlingEnabled } from 'common/hooks/useIsTxBundlingEnabled'
+import { useIsTxBundlingEnabled } from 'common/hooks/featureFlags/useIsTxBundlingEnabled'
 import { useNeedsApproval } from 'common/hooks/useNeedsApproval'
 import { withModalProvider } from 'utils/withModalProvider'
