+++ src/modules/limitOrders/hooks/useIsSafeApprovalBundle.ts
@@ -2,7 +2,7 @@ import { Currency, CurrencyAmount } from '@uniswap/sdk-core'
 
 import { Nullish } from 'types'
 
-import { useIsTxBundlingEnabled } from 'common/hooks/useIsTxBundlingEnabled'
+import { useIsTxBundlingEnabled } from 'common/hooks/featureFlags/useIsTxBundlingEnabled'
 import { useNeedsApproval } from 'common/hooks/useNeedsApproval'
 
 export function useIsSafeApprovalBundle(amount: Nullish<CurrencyAmount<Currency>>): boolean {
